refactor(scripts): rename the nine subprocess-launched scripts (ADR-0014 tier 3)
Phase 4 of the tech-debt plan. Tier 3 is the standalone scripts: never
imported by production, only launched through algo.py, so a stale path
fails loudly at the CLI rather than silently at runtime.

  Data Extractors/
    Nifty 1m 5Y Data Fetch Dhan.py     -> nifty_1m_5y_data_fetch_dhan.py
    Banknifty 1m 5Y Data Fetch Dhan.py -> banknifty_1m_5y_data_fetch_dhan.py
    Finnifty 1m 5Y Data Fetch Dhan.py  -> finnifty_1m_5y_data_fetch_dhan.py
  My Backtest Files (For Reference)/
    Nifty Renko Strategy Backtest.py       -> renko_strategy_backtest.py
    Nifty EMA Trend Strategy Backtest.py   -> ema_trend_strategy_backtest.py
    Nifty Heiken Ashi Futures 5Y Backtest.py -> heikin_ashi_futures_5y_backtest.py
    Nifty CPR Strategy Backtest.py         -> cpr_strategy_backtest.py
    Subhamoy Strategies/
      Nifty Goldmine Strategy Backtest.py     -> goldmine_strategy_backtest.py
      Nifty Money Machine Strategy Backtest.py-> money_machine_strategy_backtest.py

All nine recorded by git as pure renames. 45 references updated across
nine files: algo.py (11), the four READMEs, docs/lld/data-and-backtesting,
three test literals, and two backtests that named themselves in their own
usage docstrings.

THE PLAN EXPECTED THREE EXCLUDE LISTS TO MOVE. THEY DID NOT.

Ruff's per-file-ignores, coverage's omit, and CI's Bandit -x all name the
"My Backtest Files (For Reference)" FOLDER, not any filename. Tier 3
renames basenames only -- ADR-0014 keeps folder names deliberately -- so
every exclude still matches. Checked rather than assumed.

THE THREE DATA EXTRACTORS JOINED THE TYPE GATE; THE SIX BACKTESTS DID NOT

Same test as tier 4, opposite answers, so they are treated differently
rather than uniformly:
  * The wrappers pass mypy unmodified and are now in [tool.mypy] files.
    Their shared engine was already checked; they were outside only
    because of their names. mypy: 60 -> 63 files.
  * The backtests do NOT pass (cpr_strategy_backtest.py alone reports 6
    errors). They stay out. They are reference-only material, already
    excluded from coverage, Ruff E402 and Bandit, and typing them is not
    what this tier is for.

docs/lld/data-and-backtesting.md claimed the common module "is also the
only part under mypy (the wrappers have spaces in their names)". That is
now false in both halves, so it was rewritten rather than left to rot.

"Heiken" was normalised to "heikin" to match the rest of the repository
(heikin_ashi_strategy_logic.py, and algo.py's own --strategy heikin key,
which was already spelled that way).

The ASCII trees in the LLD kept their alignment for free: spaces became
underscores, so every name is exactly the same length as before.

ADR-0009 still names one old path in its example block. That is left
alone on purpose -- it is a superseded, point-in-time record of the world
before this change, and rewriting history inside it would make the
decision it documents unreadable.

VERIFICATION

Beyond the gates: every script path in algo.py's INDEX_SCRIPTS,
BACKTEST_SCRIPTS, BROKER_DIAGNOSTICS and the single-script constants was
parsed out of the AST and checked to exist on disk, and `algo.py backtest
--strategy renko --help` launches a renamed script end to end.

Gates: 547 master, 28 market-data-health, 1252 pytest, ruff, mypy (63
files), compileall, bandit all clean.
